hotfix : 아이템 등록시 요청받는 ItemCondition의 price를 Long으로 설정-> 메랜지지와 일치 및 아이템 등록시 메세지 수정

+++ src/main/kotlin/taeyun/malanalter/alertitem/dto/ItemCondition.kt
@@ -7,8 +7,8 @@ import kotlinx.serialization.Serializable
  */
 @Serializable
 data class ItemCondition(
-    val lowPrice: Int? = null,
-    val highPrice: Int? = null,
+    val lowPrice: Long? = null,
+    val highPrice: Long? = null,
     val str: Int? = null,
     val dex: Int? = null,
     val int: Int? = null,
